requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
httpx>=0.24.0
typing-extensions>=4.5.0